refactor: move files related to Line
